feat: rename network to chain
